feat: rename repository to storage and add interfaces
