+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -3,7 +3,8 @@ package ru.yandex.practicum.filmorate.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.model.Film;
-import ru.yandex.practicum.filmorate.repository.FilmRepository;
+import ru.yandex.practicum.filmorate.storage.FilmStorage;
+import ru.yandex.practicum.filmorate.storage.InMemoryFilmStorage;
 
 import java.util.Collection;
 import java.util.Optional;
@@ -11,17 +12,17 @@ import java.util.Optional;
 @RequiredArgsConstructor
 @Service
 public class FilmService {
-    private final FilmRepository filmRepository;
+    private final FilmStorage filmStorage;
 
     public Collection<Film> findAll() {
-        return filmRepository.findAll();
+        return filmStorage.findAll();
     }
 
     public Film save(Film film) {
-        return filmRepository.persist(film);
+        return filmStorage.persist(film);
     }
 
     public Optional<Film> update(Film film) {
-        return filmRepository.update(film);
+        return filmStorage.update(film);
     }
 }
