+++ src/main/java/ru/yandex/practicum/filmorate/storage/FilmStorage.java
@@ -1,0 +1,14 @@
+package ru.yandex.practicum.filmorate.storage;
+
+import ru.yandex.practicum.filmorate.model.Film;
+
+import java.util.Collection;
+import java.util.Optional;
+
+public interface FilmStorage {
+    Collection<Film> findAll();
+
+    Film persist(Film film);
+
+    Optional<Film> update(Film film);
+}
