+++ src/main/java/ru/yandex/practicum/filmorate/storage/InMemoryFilmStorage.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.repository;
+package ru.yandex.practicum.filmorate.storage;
 
 import org.springframework.stereotype.Component;
 import ru.yandex.practicum.filmorate.model.Film;
@@ -9,20 +9,23 @@ import java.util.Map;
 import java.util.Optional;
 
 @Component
-public class FilmRepository {
+public class InMemoryFilmStorage implements FilmStorage {
     private final Map<Long, Film> films = new HashMap<>();
     private long id = 0;
 
+    @Override
     public Collection<Film> findAll() {
         return films.values();
     }
 
+    @Override
     public Film persist(Film film) {
         film.setId(generateId());
         films.put(film.getId(), film);
         return film;
     }
 
+    @Override
     public Optional<Film> update(Film film) {
         return Optional.ofNullable(films.computeIfPresent(film.getId(), (k, v) -> film));
     }
