+++ src/main/java/ru/yandex/practicum/filmorate/storage/InMemoryUserStorage.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.repository;
+package ru.yandex.practicum.filmorate.storage;
 
 import org.springframework.stereotype.Component;
 import ru.yandex.practicum.filmorate.model.User;
@@ -9,21 +9,23 @@ import java.util.Map;
 import java.util.Optional;
 
 @Component
-public class UserRepository {
+public class InMemoryUserStorage implements UserStorage {
     private final Map<Long, User> users = new HashMap<>();
     private long id = 0;
 
-
+    @Override
     public Collection<User> findAll() {
         return users.values();
     }
 
+    @Override
     public User persist(User user) {
         user.setId(generateId());
         users.put(user.getId(), user);
         return user;
     }
 
+    @Override
     public Optional<User> update(User user) {
         return Optional.ofNullable(users.computeIfPresent(user.getId(), (k, v) -> user));
     }
