+++ src/main/java/ru/yandex/practicum/filmorate/storage/UserStorage.java
@@ -1,0 +1,14 @@
+package ru.yandex.practicum.filmorate.storage;
+
+import ru.yandex.practicum.filmorate.model.User;
+
+import java.util.Collection;
+import java.util.Optional;
+
+public interface UserStorage {
+    Collection<User> findAll();
+
+    User persist(User user);
+
+    Optional<User> update(User user);
+}
